backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2025.2
gunicorn==21.2.0